test: unify redundant HTTPProxy tests
- cert-manager_test.go: removed because its functionality is tested elsewhere
- external-dns_test.go: ditto
- network-policy_test.go: remove redundant HTTPProxy resources

All others: retry if Certificate is not issued quickly
