+++ test/suite_test.go
@@ -60,8 +60,6 @@ var _ = Describe("Test applications", func() {
 	}
 	Context("network-policy", testNetworkPolicy)
 	Context("metallb", testMetalLB)
-	Context("external-dns", testExternalDNS)
-	Context("cert-manager", testCertManager)
 	Context("contour", testContour)
 	Context("machines-endpoints", testMachinesEndpoints)
 	Context("kube-state-metrics", testKubeStateMetrics)
